Move test/generator externs into separate libs (vs aot_test code)
This makes it tractable to use Generators-with-extern-deps with the RunGen tool.
