+++ test/CMakeLists.txt
@@ -110,11 +110,12 @@ include(../HalideGenerator.cmake)
 #   halide_define_aot_test(NAME <basename>
 #                       [GENERATED_FUNCTION <function name, including namespace]
 #                       [GENERATOR_HALIDE_TARGET Halide target, if not "host"]
-#                       [GENERATOR_ARGS extra generator args])
+#                       [GENERATOR_ARGS extra generator args]
+#                       [FILTER_DEPS extra libraries to be linked with AOT output])
 function(halide_define_aot_test NAME)
   set(options OMIT_DEFAULT_GENERATOR)
   set(oneValueArgs GENERATED_FUNCTION GENERATOR_NAME GENERATOR_HALIDE_TARGET)
-  set(multiValueArgs GENERATOR_ARGS STUB_DEPS)
+  set(multiValueArgs GENERATOR_ARGS STUB_DEPS FILTER_DEPS)
   cmake_parse_arguments(args "${options}" "${oneValueArgs}" "${multiValueArgs}" ${ARGN})
 
   set(TARGET "generator_aot_${NAME}")
@@ -156,6 +157,10 @@ function(halide_define_aot_test NAME)
                            GENERATOR_ARGS "target=${args_GENERATOR_HALIDE_TARGET}" "${args_GENERATOR_ARGS}")
     halide_add_aot_library_dependency("${TARGET}" "${AOT_LIBRARY_TARGET}")
   endif()
+
+  foreach(F ${args_FILTER_DEPS})
+    target_link_libraries("${TARGET}" PRIVATE ${F})
+  endforeach()
 
 endfunction(halide_define_aot_test)
 
@@ -268,9 +273,17 @@ if (WITH_TEST_GENERATORS)
   halide_define_aot_test(user_context_insanity
                          GENERATOR_HALIDE_TARGET host-user_context)
 
+  add_library(cxx_mangling_externs 
+              "${CMAKE_CURRENT_SOURCE_DIR}/generator/cxx_mangling_externs.cpp")
+  if (NOT MSVC)
+    target_compile_options(cxx_mangling_externs PUBLIC "-std=c++11")
+  endif()
+
   halide_define_aot_test(cxx_mangling
                          GENERATOR_HALIDE_TARGET host-c_plus_plus_name_mangling
-                         GENERATED_FUNCTION HalideTest::AnotherNamespace::cxx_mangling)
+                         GENERATED_FUNCTION HalideTest::AnotherNamespace::cxx_mangling
+                         FILTER_DEPS cxx_mangling_externs)
+
   if (TARGET_PTX)
     halide_add_aot_test_dependency(cxx_mangling
                                  AOT_LIBRARY_TARGET cxx_mangling_gpu
@@ -332,9 +345,17 @@ if (WITH_TEST_GENERATORS)
                                  GENERATOR_TARGET blur2x2
                                  AOT_LIBRARY_TARGET blur2x2)
 
+  add_library(cxx_mangling_define_extern_externs 
+              "${CMAKE_CURRENT_SOURCE_DIR}/generator/cxx_mangling_define_extern_externs.cpp")
+  if (NOT MSVC)
+    target_compile_options(cxx_mangling_define_extern_externs PUBLIC "-std=c++11")
+  endif()
+  halide_add_aot_library_dependency(cxx_mangling_define_extern_externs cxx_mangling)
+
   halide_define_aot_test(cxx_mangling_define_extern
                          GENERATOR_HALIDE_TARGET host-c_plus_plus_name_mangling-user_context
-                         GENERATED_FUNCTION HalideTest::cxx_mangling_define_extern)
+                         GENERATED_FUNCTION HalideTest::cxx_mangling_define_extern
+                         FILTER_DEPS cxx_mangling_externs cxx_mangling_define_extern_externs)
   # The cxx_mangling library was already defined implicitly, above,
   # so just add a dependency on it
   halide_add_aot_library_dependency(generator_aot_cxx_mangling_define_extern cxx_mangling)
